requirements.txt
boto3
s3fs
pandas
jupyter
notebook
xarray
cdsapi
netCDF4
scikit-learn
matplotlib
seaborn
pyarrow
python-dotenv
npm
aws-cdk-lib
constructs>=10.0.0
rasterio
beautifulsoup4
requests